Bug 706, Bug 707: revamped the way Purr handles watched directories. We have a proper list widget now, plus buttons to add and remove directories. "Pounce" and "watch" is now enabled on a per-directory basis via a tri-state checkbox.  Added a starup wizard to handle connecting to the appropriate purrlog.
In Calico, added a ParmTables module, for high-level access to parmtables. 
Added a calico-parmgroomer.py script to use this to look at parms.  This
forced some changes in the kernel, specifically related to handling of domains
and parms. Also changed mequtils.cc, so that axis IDs are returned as all-lowercase
strings.